+++ src/id3/v2/write/mod.rs
@@ -12,8 +12,23 @@ use crate::probe::Probe;
 
 use std::fs::File;
 use std::io::{Cursor, Read, Seek, SeekFrom, Write};
+use std::ops::Not;
 
 use byteorder::{BigEndian, LittleEndian, WriteBytesExt};
+
+// In the very rare chance someone wants to write a CRC in their extended header
+static CRC_32_TABLE: once_cell::sync::Lazy<[u32; 256]> = once_cell::sync::Lazy::new(|| {
+	let mut crc32_table = [0; 256];
+
+	for n in 0..256 {
+		crc32_table[n as usize] = (0..8).fold(n as u32, |acc, _| match acc & 1 {
+			1 => 0xEDB8_8320 ^ (acc >> 1),
+			_ => acc >> 1,
+		});
+	}
+
+	crc32_table
+});
 
 #[allow(clippy::shadow_unrelated)]
 pub(crate) fn write_id3v2<'a, I: Iterator<Item = FrameRef<'a>> + 'a>(
@@ -67,7 +82,11 @@ pub(super) fn create_tag<'a, I: Iterator<Item = FrameRef<'a>> + 'a>(
 	}
 
 	let has_footer = tag.flags.footer;
-	let mut id3v2 = create_tag_header(tag.flags)?;
+	let needs_crc = tag.flags.crc;
+	#[cfg(feature = "id3v2_restrictions")]
+	let has_restrictions = tag.flags.restrictions.0;
+
+	let (mut id3v2, extended_header_len) = create_tag_header(tag.flags)?;
 	let header_len = id3v2.get_ref().len();
 
 	// Write the items
@@ -77,7 +96,31 @@ pub(super) fn create_tag<'a, I: Iterator<Item = FrameRef<'a>> + 'a>(
 
 	// Go back to the start and write the final size
 	id3v2.seek(SeekFrom::Start(6))?;
-	id3v2.write_u32::<BigEndian>(synch_u32(len as u32)?)?;
+	id3v2.write_u32::<BigEndian>(synch_u32(extended_header_len + len as u32)?)?;
+
+	if needs_crc {
+		// The CRC is calculated on all the data between the header and footer
+		#[allow(unused_mut)]
+		// Past the CRC
+		let mut content_start_idx = 22;
+
+		#[cfg(feature = "id3v2_restrictions")]
+		if has_restrictions {
+			content_start_idx += 3;
+		}
+
+		// Skip 16 bytes
+		//
+		// Normal ID3v2 header (10)
+		// Extended header (6)
+		id3v2.seek(SeekFrom::Start(16))?;
+
+		let tag_contents = &id3v2.get_ref()[content_start_idx..];
+		let encoded_crc = calculate_crc(tag_contents);
+
+		id3v2.write_u8(5)?;
+		id3v2.write_all(&encoded_crc)?;
+	}
 
 	if has_footer {
 		id3v2.seek(SeekFrom::Start(3))?;
@@ -94,7 +137,7 @@ pub(super) fn create_tag<'a, I: Iterator<Item = FrameRef<'a>> + 'a>(
 	Ok(id3v2.into_inner())
 }
 
-fn create_tag_header(flags: Id3v2TagFlags) -> Result<Cursor<Vec<u8>>> {
+fn create_tag_header(flags: Id3v2TagFlags) -> Result<(Cursor<Vec<u8>>, u32)> {
 	let mut header = Cursor::new(Vec::new());
 
 	header.write_all(&[b'I', b'D', b'3'])?;
@@ -125,8 +168,7 @@ fn create_tag_header(flags: Id3v2TagFlags) -> Result<Cursor<Vec<u8>>> {
 	header.write_u8(tag_flags)?;
 	header.write_u32::<BigEndian>(0)?;
 
-	// TODO
-	#[allow(unused_mut)]
+	let mut extended_header_size = 0;
 	if extended_header {
 		// Structure of extended header:
 		//
@@ -138,20 +180,20 @@ fn create_tag_header(flags: Id3v2TagFlags) -> Result<Cursor<Vec<u8>>> {
 		// Start with a zeroed header
 		header.write_all(&[0; 6])?;
 
-		let mut size = 6_u32;
+		extended_header_size = 6_u32;
 		let mut ext_flags = 0_u8;
 
 		if flags.crc {
-			// ext_flags |= 0x20;
-			// size += 5;
-			//
-			// header.write_all(&[5, 0, 0, 0, 0, 0])?;
+			ext_flags |= 0x20;
+			extended_header_size += 6;
+
+			header.write_all(&[0; 6])?;
 		}
 
 		#[cfg(feature = "id3v2_restrictions")]
 		if flags.restrictions.0 {
 			ext_flags |= 0x10;
-			size += 2;
+			extended_header_size += 2;
 
 			header.write_u8(1)?;
 			header.write_u8(flags.restrictions.1.as_bytes())?;
@@ -160,10 +202,73 @@ fn create_tag_header(flags: Id3v2TagFlags) -> Result<Cursor<Vec<u8>>> {
 		header.seek(SeekFrom::Start(10))?;
 
 		// Seek back and write the actual values
-		header.write_u32::<BigEndian>(synch_u32(size)?)?;
+		header.write_u32::<BigEndian>(synch_u32(extended_header_size)?)?;
 		header.write_u8(1)?;
 		header.write_u8(ext_flags)?;
+
+		header.seek(SeekFrom::End(0))?;
 	}
 
-	Ok(header)
+	Ok((header, extended_header_size))
+}
+
+// https://github.com/rstemmer/id3edit/blob/0246f3dc1a7a80a64461eeeb7b9ee88379003eb1/encoding/crc.c#L6:6
+fn calculate_crc(content: &[u8]) -> [u8; 5] {
+	let crc: u32 = content
+		.iter()
+		.fold(!0, |crc, octet| {
+			(crc >> 8) ^ CRC_32_TABLE[(((crc & 0xFF) ^ u32::from(*octet)) & 0xFF) as usize]
+		})
+		.not();
+
+	// The CRC-32 is stored as an 35 bit synchsafe integer, leaving the upper
+	// four bits always zeroed.
+	let mut encoded_crc = [0; 5];
+	let mut b;
+
+	#[allow(clippy::needless_range_loop)]
+	for i in 0..5 {
+		b = (crc >> ((4 - i) * 7)) as u8;
+		b &= 0x7F;
+		encoded_crc[i] = b;
+	}
+
+	encoded_crc
+}
+
+#[cfg(test)]
+mod tests {
+	use crate::id3::v2::{Id3v2Tag, Id3v2TagFlags};
+	use crate::{Accessor, TagExt};
+
+	#[test]
+	fn id3v2_write_crc32() {
+		let mut tag = Id3v2Tag::default();
+		tag.set_artist(String::from("Foo artist"));
+
+		let flags = Id3v2TagFlags {
+			crc: true,
+			..Id3v2TagFlags::default()
+		};
+		tag.set_flags(flags);
+
+		let mut writer = Vec::new();
+		tag.dump_to(&mut writer).unwrap();
+
+		let crc_content = &writer[16..22];
+		assert_eq!(crc_content, &[5, 0x06, 0x35, 0x69, 0x7D, 0x14]);
+
+		// Get rid of the size byte
+		let crc_content = &crc_content[1..];
+		let mut unsynch_crc = 0;
+
+		#[allow(clippy::needless_range_loop)]
+		for i in 0..5 {
+			let mut b = crc_content[i];
+			b &= 0x7F;
+			unsynch_crc |= u32::from(b) << ((4 - i) * 7);
+		}
+
+		assert_eq!(unsynch_crc, 0x66BA_7E94);
+	}
 }