+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     This provides additional information for the type of audio being dealt with.
 - `TagExt::clear`
   - This allows tags to be cleared of any items or pictures, while retaining any flags (if applicable)
+- **ID3v2**: Respect `Id3v2TagFlags::crc` when writing
+  - Previously, this flag was ignored, but it will now calculate a CRC for the extended header
 
 ### Changed
 - **MP4**: Sample rates are now retrieved from the [audio specific config](https://wiki.multimedia.cx/index.php?title=MPEG-4_Audio#Audio_Specific_Config) (if possible).